// src/protos/gen/scheduler/GetDoctorReviews.ts
// Original file: src/protos/review.proto


export interface GetDoctorReviews {
  'access_token'?: (string);
  'request_query'?: ({[key: string]: string});
  'doctorId'?: (string);
}

export interface GetDoctorReviews__Output {
  'access_token': (string);
  'request_query': ({[key: string]: string});
  'doctorId': (string);
}